Extracted deleteTaskEvent to its own Helper-class to keep UI-classes clean.
